Add unit tests and refactor
Create explicit tests for refactored folder path validation. Unsure of
the use of bool returning function with both an `out` and optional `bool`
parameter. A bit of a code smell that may need to be tackled hereafter.

Initially used Rename refactoring as the example for AddComponentViewModel.
After further review decided to refactor and check for name validity while
including ProjectID permitting only non-colliding names to be used and
preventing the need for a subsequent dialog to be shown/confirmed.